1. Extract Const Values 2. ShipDesign.java consts 3. to separate const values from ShipDesign
