+++ src/test/java/org/openRealmOfStars/player/ship/ShipTest.java
@@ -3,6 +3,7 @@ package org.openRealmOfStars.player.ship;
 import org.junit.Test;
 import org.junit.experimental.categories.Category;
 import org.openRealmOfStars.player.SpaceRace.SpaceRace;
+import org.openRealmOfStars.player.ship.shipdesign.ShipDesign;
 
 import static org.junit.Assert.*;
